feat(est): Add Estimate-related methods, types
parts of the response are integer-like in some cases and strings in
others. Such as the "Minutes" field in Estimates, which will say
"Leaving" instead of "0". So, add a special type to parse that into int.

fix(api): Add input validation for platform, direction params in estimates

The Bart API does provide this error information, however it's in XML
instead of JSON. For now I figured, it'd be easier to catch this on
client than it would be to parse XML, and possibly convert to JSON.