feat(sched/routesched): Add route schedule request method & response types
Add stub validations. I'm beginning to think that it'd be better to not
validate input params, and instead let the request go through so that
the BART api can relay the error message. Certain combinations of inputs
give this API route some trouble and trying to mimic the logic here does
not seem like a good idea. The problem is, error messages come back as
XML, so I'd need to convert that to JSON (BART says that JSON version of
API is in beta).