feat(est): Add EstimateParams type
Makes requesting a real time estimate less clunky. Function signatures
where arguments are all positional and of the same type is suboptimal,
as is the case in RequestETD method. Add a new method, RequestEstimate,
rather than replace the RequestETD method in the slight chance that
there is some application out there that relies on it. Backwards
compatibility is pretty important in golang.